Fix localisation handling for templates
The previous solution only worked in testing because of misconfiguration …

Apparently TYPO3 Views have no default handling for the language path in analogy to those for templates, partials etc.

It is hard to access the »view« array of TypoScript settings from the code, so we cannot access it in the template.

The ugly solution is to move this TypoScript setting into the »settings« array, away from the other path settings for the templates, partials etc.

+ Adapt templates to the new setup