germania-sacra: more aggressive highlighting
This is technically wrong but seems to give better results.
E.g. when searching for »Augustiner«, some results only have »Augustiner«
as orden but their title only contains »Augustinerchorherrenstift« which
does not match. It should not match that word but the result list looks
strange when it isn’t highlighted.