make dataFields and highlight setup options configurable per action
Turn the settings fields »dataFields« and »highlight« into an array with potential fields »default«, »index«, »detail«, »data«, »suggest«. The extension will use the default setting and add the setting for the current action to it.

This can reduce the data loaded from Solr (only the fields used in the list loaded with the / highlighting limited to the fields actually displayed for the index action).

Also set up the new configuration style in the example projects.