Only handle the first selection event for the histogram facet
Strangely two events with different selection
ranges can occur, with the first one containing
the correct range (as shown in our tooltip) and
the second one containing a smaller range.

To ensure the new facet is consistent with the
tooltip, only handle this event once, so the
second call does not »win« the race to setting
the new window.location.