find.js: attempt to fix some off-by-one errors in histogram selections
We don’t have a proper JS testing setup but this
test would not give any error output:

roundedRangeTest = function () {
    [
    {range: {from: 0, to: 0}, barWidth: 1, expected: {from: 0, to: 1}},
    {range: {from: 0, to: 0.5}, barWidth: 1, expected: {from: 0, to: 1}},
    {range: {from: 0, to: 1}, barWidth: 1, expected: {from: 0, to: 2}},
    {range: {from: 0, to: 0.5}, barWidth: 2, expected: {from: 0, to: 2}},
    {range: {from: 0, to: 1}, barWidth: 2, expected: {from: 0, to: 2}},
    {range: {from: 0.1, to: 0.1}, barWidth: 1, expected: {from: 0, to: 1}},
    {range: {from: 0.2, to: 1.1}, barWidth: 1, expected: {from: 0, to: 2}},
    {range: {from: 0, to: 0}, barWidth: 1, expected: {from: 0, to: 1}},
    ].forEach(function (testData) {
        facetConfig.barWidth = testData.barWidth;
        var result = roundedRange(testData.range);
        if (result.from !== testData.expected.from || result.to !== testData.expected.to) {
        console.log('incorrect result ' + JSON.stringify(result) + ' for case ' + JSON.stringify(testData));
        }
    });
} 

The focus was on making the selection and its
tooltip exactly correspond to the selected range
for the next query. The code continues to contain
a number of intentional »off-by-one« values when
it comes to drawing the selection as even a single
value selection needs to have a non-zero width
to be visible.

In addition, this change avoids reloading the page
if the user-created selection is already in the
active facets. Not only to have fewer page loads
but also because in some setups a different order
of JS execution could cause infinite reload loops.